// frontend/src/types/orchestrator.ts
export type JobState = 
  | 'Pending'
  | 'Running'
  | 'Successful'
  | 'Faulted'
  | 'Canceled'
  | 'Stopping'
  | 'Terminating'
  | 'Resumed';
// ...
export interface CountStats {
  title: string;
  count: number;
}

export interface LicenseStatsModel {
  Date: string;
  Available: number;
  Used: number;
  Total: number;
}

export interface ConsumptionLicenseStatsModel {
  Date: string;
  Consumption: number;
}

export interface ODataResponse<T> {
  '@odata.context'?: string;
  value: T[];
  '@odata.count'?: number;
}

export interface JobFilters {
  folderId?: number;
  processKey?: string;
  releaseKey?: string;
  robotId?: number;
  state?: JobState | JobState[];
  startDate?: string;
  endDate?: string;
  machineId?: number;
}